--- a/PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/webapp/WebAppComboBox.java
+++ b/PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/webapp/WebAppComboBox.java
@@ -6,14 +6,11 @@
 package com.microsoft.azure.toolkit.intellij.webapp;
 
 import com.intellij.openapi.project.Project;
-import com.microsoft.azure.management.appservice.WebApp;
 import com.microsoft.azure.toolkit.intellij.appservice.AppServiceComboBox;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.webapp.WebAppService;
 import com.microsoft.azuretools.azurecommons.helpers.NotNull;
-import com.microsoft.azuretools.core.mvp.model.ResourceEx;
 import com.microsoft.azuretools.core.mvp.model.webapp.AzureWebAppMvpModel;
-import com.microsoft.azuretools.utils.WebAppUtils;
 import com.microsoft.tooling.msservices.components.DefaultLoader;
 
 import java.util.List;
@@ -48,11 +45,9 @@ public class WebAppComboBox extends AppServiceComboBox<WebAppComboBoxModel> {
         type = AzureOperation.Type.SERVICE
     )
     protected List<WebAppComboBoxModel> loadItems() throws Exception {
-        final List<ResourceEx<WebApp>> webApps = AzureWebAppMvpModel.getInstance().listAllWebApps(false);
-        return webApps.stream()
-                      .filter(resource -> WebAppUtils.isJavaWebApp(resource.getResource()))
-                      .sorted((a, b) -> a.getResource().name().compareToIgnoreCase(b.getResource().name()))
-                      .map(webAppResourceEx -> new WebAppComboBoxModel(webAppResourceEx))
-                      .collect(Collectors.toList());
+        return AzureWebAppMvpModel.getInstance().listJavaWebApps(false).parallelStream()
+            .map(WebAppComboBoxModel::new)
+            .sorted((app1, app2) -> app1.getAppName().compareToIgnoreCase(app2.getAppName()))
+            .collect(Collectors.toList());
     }
 }