--- a/PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/webapp/WebAppComboBox.java
+++ b/PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/webapp/WebAppComboBox.java
@@ -8,6 +8,7 @@ package com.microsoft.azure.toolkit.intellij.webapp;
 import com.intellij.openapi.project.Project;
 import com.microsoft.azure.management.appservice.WebApp;
 import com.microsoft.azure.toolkit.intellij.appservice.AppServiceComboBox;
+import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.webapp.WebAppService;
 import com.microsoft.azuretools.azurecommons.helpers.NotNull;
 import com.microsoft.azuretools.core.mvp.model.ResourceEx;
@@ -42,6 +43,10 @@ public class WebAppComboBox extends AppServiceComboBox<WebAppComboBoxModel> {
 
     @NotNull
     @Override
+    @AzureOperation(
+        name = "webapp.list.detail|subscription|selected",
+        type = AzureOperation.Type.SERVICE
+    )
     protected List<WebAppComboBoxModel> loadItems() throws Exception {
         final List<ResourceEx<WebApp>> webApps = AzureWebAppMvpModel.getInstance().listAllWebApps(false);
         return webApps.stream()