… Utils/azuretools-core/resources/com/microsoft/azure/toolkit/operation/titles.properties