a proof of concept about how to improve performance
listing webapps takes `>80s` during deploying webapp for the first time

I improved the performance with these machanisms:
* sequential -> parallel: e.g. use parallel streams : `>80s` -> `~20s` on my laptop(8 cores?)
* cache + warmup: 
  * cache only helps later calls but not the first. 
  * Async cache warmup decrease timecost from `>80s` to `<1s`. but slows a bit the ide starting up according to my observation.
* lazy-load: not used in this poc.